backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
psycopg2-binary==2.9.10
python-dotenv==1.1.0
Werkzeug==3.1.3
gunicorn==23.0.0